trident-acl-agent: restructure src into core/annotations/omahaonly modules
Pure reorganization, no logic changes. Groups shared infrastructure
under core/ (config, error, id, version, trident client, nebraska
client), the Kubernetes-annotation-driven default flow under
annotations/ (protocol, k8s, state, orchestrator), and the legacy
one-shot flow under omahaonly/.

- core/trident/client.rs (was trident.rs) and core/trident/mock.rs
  (was mock_tridentd.rs) avoid path stutter under core::trident::.
- annotations/protocol.rs (was annotations.rs) frees up the
  annotations module name for the directory; annotations/mod.rs
  glob re-exports it so crate::annotations::* call sites are
  unchanged.
- core/mod.rs keeps config/error/id/nebraska/trident/version nested
  (not flattened) since config.rs in particular has a large public
  surface.

152 unit tests + 1 doctest pass; clippy and fmt clean.
