+++ crates/trident-acl-agent/src/core/trident/mod.rs
@@ -1,0 +1,8 @@
+//! gRPC client for talking to `tridentd`, plus its in-process mock used only
+//! by unit tests.
+
+mod client;
+pub use client::*;
+
+#[cfg(test)]
+pub mod mock;
