+++ crates/trident-acl-agent/src/omahaonly/mod.rs
@@ -1,0 +1,68 @@
+//! The historical one-shot Omaha flow, preserved as an explicit opt-out from
+//! the default annotation-driven protocol (see
+//! [`crate::core::config::GoalSource`]).
+
+use anyhow::Context;
+use semver::Version;
+
+use crate::core::{
+    config::AgentConfig,
+    nebraska::{CheckOutcome, Client},
+    trident::TridentClient,
+    version,
+};
+use crate::IdSource;
+
+/// Historical one-shot flow: query the Nebraska/Omaha server at
+/// `config.nebraska.endpoint` once, and if an update is offered, call
+/// tridentd's combined `Update()` RPC once and exit. No Kubernetes/annotation
+/// involvement.
+pub async fn run_omaha_only(config: &AgentConfig) -> Result<(), anyhow::Error> {
+    let endpoint = config.nebraska.endpoint.clone().ok_or_else(|| {
+        anyhow::anyhow!("no Nebraska endpoint configured: set TRIDENT_ACL_AGENT_NEBRASKA_ENDPOINT")
+    })?;
+
+    // Client::check_for_update() is a blocking call (reqwest::blocking under
+    // the hood, see nebraska::transport) - calling it directly from this
+    // async fn can panic ("Cannot drop a runtime in a context where blocking
+    // is not allowed") because reqwest::blocking spins up its own inner
+    // Tokio runtime per call, which isn't safe to tear down from inside an
+    // already-running async task. Run it on a dedicated blocking thread.
+    let app_id = config.nebraska.app_id.clone();
+    let track = config.nebraska.track.clone();
+    let machine_id = crate::build_machine_id(IdSource::MachineIdHashed)?;
+    let current_version_raw = version::current_active_version()?;
+    let current_version = Version::parse(&current_version_raw).unwrap_or_else(|err| {
+        log::warn!(
+            "current version {current_version_raw:?} is not valid semver ({err}); reporting 0.0.0 to Nebraska"
+        );
+        Version::new(0, 0, 0)
+    });
+    let outcome = tokio::task::spawn_blocking(move || {
+        let client = Client::new(endpoint, app_id, track, machine_id);
+        client.check_for_update(&current_version)
+    })
+    .await
+    .context("Nebraska query task panicked")?
+    .map_err(|err| anyhow::anyhow!("Nebraska query failed: {err}"))?;
+
+    match outcome {
+        CheckOutcome::UpToDate | CheckOutcome::UpdateInProgress => {
+            log::debug!("No update available from Nebraska");
+            Ok(())
+        }
+        CheckOutcome::UpdateAvailable(offer) => {
+            log::info!("Triggering one-shot Omaha update to {}", offer.version);
+            let mut client = TridentClient::connect(&config.trident.socket).await?;
+            let combined_timeout =
+                config.orchestration.stage_timeout + config.orchestration.finalize_timeout;
+            // Integrity of the downloaded image is verified by Trident itself
+            // via the image's own COSI metadata, so the Nebraska-reported hash
+            // (offer.primary.hash) is not passed here.
+            client
+                .update(&offer.primary.url, None, combined_timeout)
+                .await?;
+            Ok(())
+        }
+    }
+}
